migration bug fix

+++ database/migrations/2020_05_19_190713_create_users_table.php
@@ -19,8 +19,8 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->unsignedInteger('role_id')->nullable();
-            $table->unsignedInteger('gender_id')->nullable();
+            $table->unsignedBigInteger('role_id')->nullable();
+            $table->unsignedBigInteger('gender_id')->nullable();
             $table->foreign('role_id')->references('id')->on('roles');
             $table->foreign('gender_id')->references('id')->on('genders');
             $table->rememberToken();
